+++ internal/provider/provider.go
@@ -157,6 +157,7 @@ func (p *scalrProvider) Resources(_ context.Context) []func() resource.Resource 
 
 func (p *scalrProvider) DataSources(_ context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
+		newProviderConfigurationDataSource,
 		newTagDataSource,
 	}
 }