+++ internal/provider/provider_sdk.go
@@ -47,7 +47,6 @@ func Provider(v string) *schema.Provider {
 			"scalr_module_version":           dataSourceModuleVersion(),
 			"scalr_module_versions":          dataSourceModuleVersions(),
 			"scalr_policy_group":             dataSourceScalrPolicyGroup(),
-			"scalr_provider_configuration":   dataSourceScalrProviderConfiguration(),
 			"scalr_provider_configurations":  dataSourceScalrProviderConfigurations(),
 			"scalr_role":                     dataSourceScalrRole(),
 			"scalr_service_account":          dataSourceScalrServiceAccount(),